Moved sendError to backend utils
