+++ src/server/utils.ts
@@ -1,5 +1,5 @@
 import { Response } from 'express';
-import { toMsg } from './utils';
+import { toMsg } from '../common/utils';
 
 export function sendError(res: Response, error: string, status: number = 500) {
     res.status(status).send({ error: toMsg(error)});
